+++ CHANGELOG.md
@@ -1,1 +1,38 @@
-# pkg_multimedia## Changelog### 1.1.3 - 14 january 2015* Added Changelog* Updated install script### 1.1.2 - 12 november 2014* Updated keywords* New taxonomy structure### 1.1.1 - 31 august 2014* Updated documentation### 1.1.0 - 11 august 2014* Added Vimeo support### 1.0.2 - 7 august 2014* Added composer support### 1.0.1 - 24 july 2014* Minor typo fixes### 1.0.0 - 25 june 2014* Initial launch
+# pkg_multimedia
+
+## Changelog
+
+### 1.2.0 - 19 januray 2015
+
+* Added 'translatable' sources
+* Added source edit screen
+
+### 1.1.3 - 14 january 2015
+
+* Added Changelog
+* Updated install script
+
+### 1.1.2 - 12 november 2014
+
+* Updated keywords
+* New taxonomy structure
+
+### 1.1.1 - 31 august 2014
+
+* Updated documentation
+
+### 1.1.0 - 11 august 2014
+
+* Added Vimeo support
+
+### 1.0.2 - 7 august 2014
+
+* Added composer support
+
+### 1.0.1 - 24 july 2014
+
+* Minor typo fixes
+
+### 1.0.0 - 25 june 2014
+
+* Initial launch